afiliacion-amazon/backend/microservicios/api-paapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
paapi5-python-sdk==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.25.1
